更改评分

+++ src/main/java/com/gdut/pandora/domain/Topic.java
@@ -30,7 +30,7 @@ public class Topic {
 
     private Boolean isAnonymous;
 
-    private Long likeNum;
+    private Integer likeNum;
 
     private Long createTime;
